pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "djangorestframework-stubs"
version = "0.1.0"
description = "Type stubs for Django REST Framework"
readme = "README.md"
requires-python = ">=3.7"
license = "MIT"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
]
dependencies = [
    "djangorestframework>=3.12",
]

[project.urls]
"Homepage" = "https://github.com/foxman2/rest_framework-stubs"
"Bug Tracker" = "https://github.com/foxman2/rest_framework-stubs/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["rest_framework*"]

[tool.setuptools.package-data]
"rest_framework" = ["py.typed", "*.pyi"]
"rest_framework.authtoken" = ["*.pyi"]
"rest_framework.authtoken.management" = ["*.pyi"]
"rest_framework.authtoken.management.commands" = ["*.pyi"]
"rest_framework.management" = ["*.pyi"]
"rest_framework.management.commands" = ["*.pyi"]
"rest_framework.schemas" = ["*.pyi"]
"rest_framework.templatetags" = ["*.pyi"]
"rest_framework.utils" = ["*.pyi"]
